# Optimize EvaluateMemberExpression when targeting static member