Member expression invocation is on hot path with static object properties and allocates a lot of literals, we can simplify it to work without literal expression evaluation.

## UncacheableExpressionsBenchmark

### Before (dev branch)

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.094 s | 0.0049 s | 0.0071 s | 243554.1667 | 15075.0000 | 997.15 MB |

### After (this pull request)

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.081 s | 0.0112 s | 0.0167 s | 202687.5000 | 55312.5000 |  903.2 MB |